requirements.txt
typer
i18nice
i18nice[YAML]
typing_extensions>=4.7
